Rename: UserCommand 위치 변경
