+++ user-api/src/main/java/com/biengual/userapi/user/presentation/UserDtoMapper.java
@@ -1,7 +1,7 @@
 package com.biengual.userapi.user.presentation;
 
 import com.biengual.userapi.oauth2.domain.info.OAuth2UserPrincipal;
-import com.biengual.userapi.user.application.UserCommand;
+import com.biengual.userapi.user.domain.UserCommand;
 import org.mapstruct.InjectionStrategy;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
@@ -19,6 +19,7 @@ import org.mapstruct.ReportingPolicy;
 )
 public interface UserDtoMapper {
 
+    @Mapping(target = "username", source = "request.username")
     @Mapping(target = "email", source = "principal.email")
     UserCommand.UpdateMyInfo of(UserRequestDto.UpdateMyInfo request, OAuth2UserPrincipal principal);
 }
